+++ src/App/LinkBaseExtensionPyImp.cpp
@@ -238,7 +238,7 @@ PyObject* LinkBaseExtensionPy::setLink(PyObject *_args)
             PyObject *key, *value;
             Py_ssize_t pos = 0;
             while(PyDict_Next(pcObj, &pos, &key, &value))
-                parseLink(ext,Py::Int(key),value);
+                parseLink(ext,Py::Long(key),value);
         }else if(PySequence_Check(pcObj)) {
             ext->setLink(-1,nullptr);
             Py::Sequence seq(pcObj);
